requirements.txt
setuptools==0.7.3
sn1crypto==0.23.0
beautifulsoup4==4.6.0
certifi==2017.7.27.1
cffi==1.11.2
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.1.2
Django==1.11.2
django-cors-headers==2.1.0
django-filter==1.0.4
django-pusherable==0.2.0
django-rest-swagger==2.1.2
djangorestframework==3.7.1
drf-openapi==1.0.0
google-api-python-client==1.6.4
httplib2==0.10.3
idna==2.6
itypes==1.1.0
Jinja2==2.9.6
json-tricks==3.8.0
lxml==4.1.0
Markdown==2.6.9
MarkupSafe==1.0
ndg-httpsclient==0.4.3
oauth2client==4.1.2
openapi-codec==1.3.2
passlib==1.7.1
psycopg2==2.7.3.1
pusher==1.7.2
pyasn1==0.3.7
pyasn1-modules==0.1.5
pycparser==2.18
pyOpenSSL==17.3.0
pytz==2017.2
requests==2.18.4
rsa==3.4.2
simplejson==3.11.1
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
